requirements.txt
PyYaml
python-gitlab
